agentic_ai_backend/agents/orchestrators/pyproject.toml
[project]
name = "orchestrators"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "agent-framework-core>=1.0.0b251211",
 "azure-core>=1.37.0",
 "azure-search-documents>=11.6.0",
 "aiohttp>=3.9.0",
 "python-dotenv>=1.0.0",
]
